models: move ErrBranchNotExist to errors package
