+++ models/errors/repo.go
@@ -72,3 +72,16 @@ func IsBranchAlreadyExists(err error) bool {
 func (err BranchAlreadyExists) Error() string {
 	return fmt.Sprintf("branch already exists [name: %s]", err.Name)
 }
+
+type ErrBranchNotExist struct {
+	Name string
+}
+
+func IsErrBranchNotExist(err error) bool {
+	_, ok := err.(ErrBranchNotExist)
+	return ok
+}
+
+func (err ErrBranchNotExist) Error() string {
+	return fmt.Sprintf("branch does not exist [name: %s]", err.Name)
+}
